+++ src/hooks/validators/index.ts
@@ -1,5 +1,6 @@
 export * from './numberUtilities'
 export * from './getStringValidator'
 export * from './getGenericValidator'
+export * from './getArrayValidator'
 export * from './getError'
 export * from './types'