+++ src/utils/inputTypes.ts
@@ -1,0 +1,21 @@
+export enum InputTypes {
+  generic = 'generic',
+  radio = 'radio',
+  select = 'select',
+  input = 'input',
+  textArea = 'textArea',
+  checkbox = 'checkbox',
+  fieldArray = 'array',
+}
+
+export enum UITypes {
+  default = 'default',
+  radio = 'radio',
+  select = 'select',
+  input = 'input',
+  hidden = 'hidden',
+  password = 'password',
+  textArea = 'textArea',
+  checkbox = 'checkbox',
+  array = 'array',
+}